requirements.txt
matplotlib==3.5.2
numpy==1.21.6
pandas==1.1.5
pymongo==4.1.1
python-dotenv==0.20.0
tweepy==4.8.0
vaderSentiment==3.3.2
zemberek_python==0.1.2
